+++ tools/core-build-tasks/src/index.test.ts
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
 import { describe, expect, it } from 'vitest';
 
 describe('Remove when tests are added', () => {