+++ tools/core-build-tasks/src/tasks/index.ts
@@ -1,4 +1,14 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
 export * from './api-extractor';
+export * from './bundle';
 export * from './clean';
+export * from './cleanCollateral';
+export * from './copy';
 export * from './coreLint';
+export * from './helpers';
 export * from './vitest';
+export * from './updateWorld';
+export * from './watch';
+export * from './zip';